Inline HTTPClientRequestPartWithBaggage type
