+++ Sources/NIOInstrumentation/BaggageContextOutboundHTTPHandler.swift
@@ -26,3 +26,13 @@ public final class BaggageContextOutboundHTTPHandler: ChannelOutboundHandler {
         context.write(self.wrapOutboundOut(.head(head)), promise: promise)
     }
 }
+
+public struct HTTPClientRequestPartWithBaggage {
+    public let requestPart: HTTPClientRequestPart
+    public let baggage: BaggageContext
+
+    public init(requestPart: HTTPClientRequestPart, baggage: BaggageContext) {
+        self.requestPart = requestPart
+        self.baggage = baggage
+    }
+}
