--- Sources/NIOInstrumentation/HTTPClientRequestPartWithBaggage.swift
@@ -1,12 +1,0 @@
-import BaggageContext
-import NIOHTTP1
-
-public struct HTTPClientRequestPartWithBaggage {
-    public let requestPart: HTTPClientRequestPart
-    public let baggage: BaggageContext
-
-    public init(requestPart: HTTPClientRequestPart, baggage: BaggageContext) {
-        self.requestPart = requestPart
-        self.baggage = baggage
-    }
-}
